DOCKER/requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
altair==4.2.2
async-timeout==4.0.3
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
beautifulsoup4==4.12.2
blinker==1.4
blobfile==2.0.2
cachetools==5.3.1
certifi==2020.6.20
chardet==4.0.0
charset-normalizer==3.2.0
click==8.0.3
cloud-init==23.2.2
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.0
cryptography==3.4.8
cycler==0.11.0
dataclasses-json==0.5.14
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.1
entrypoints==0.4
filelock==3.12.2
fonttools==4.42.1
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
greenlet==2.0.2
httplib2==0.20.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
jeepney==0.7.1
Jinja2==3.0.3
joblib==1.3.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
kiwisolver==1.4.4
langchain==0.0.266
langsmith==0.0.26
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
llama-index==0.8.8
loguru==0.7.0
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.0.1
marshmallow==3.20.1
matplotlib==3.6.3
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.0.4
mypy-extensions==1.0.0
nest-asyncio==1.5.7
netifaces==0.11.0
numexpr==2.8.5
numpy==1.25.2
oauthlib==3.2.0
openai==0.27.9
openapi-schema-pydantic==1.2.4
packaging==23.1
pandas==2.0.3
pexpect==4.8.0
Pillow==9.5.0
plotly==5.12.0
protobuf==4.24.1
ptyprocess==0.7.0
pyarrow==13.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
pycryptodomex==3.18.0
pydantic==1.10.12
pydeck==0.8.0
Pygments==2.16.1
PyGObject==3.42.1
PyHamcrest==2.0.2
PyJWT==2.3.0
Pympler==1.0.1
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyparted==3.11.7
pypdf==3.15.2
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu2
python-dateutil==2.8.2
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.0
python-magic==0.4.24
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
regex==2023.8.8
requests==2.31.0
rich==13.5.2
scikit-learn==1.2.0
scipy==1.10.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
smmap==5.0.0
sos==4.5.6
soupsieve==2.4.1
SQLAlchemy==2.0.20
ssh-import-id==5.11
streamlit==1.25.0
systemd-python==234
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.3.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
tqdm==4.66.1
Twisted==22.1.0
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
validators==0.21.2
wadllib==1.3.6
WALinuxAgent==2.2.46
watchdog==3.0.0
yarl==1.9.2
zipp==1.0.0
zope.interface==5.4.0
